+++ examples/provider/provider.tf
@@ -2,7 +2,7 @@ terraform {
   required_providers {
     pingone = {
       source  = "pingidentity/pingone"
-      version = "~> 0.26"
+      version = "~> 0.27"
     }
   }
 }
@@ -12,8 +12,6 @@ provider "pingone" {
   client_secret  = var.client_secret
   environment_id = var.environment_id
   region         = var.region
-
-  force_delete_production_type = false
 }
 
 resource "pingone_environment" "my_environment" {