+++ examples/resources/pingone_notification_policy/resource-environment.tf
@@ -5,7 +5,7 @@ resource "pingone_environment" "my_environment" {
 resource "pingone_notification_policy" "environment" {
   environment_id = pingone_environment.my_environment.id
 
-  name = "Environment Quota SMS, Voice and Email"
+  name = "Environment Quota SMS Voice and Email"
 
   quota {
     type             = "ENVIRONMENT"