+++ examples/resources/pingone_notification_policy/resource-user.tf
@@ -5,7 +5,7 @@ resource "pingone_environment" "my_environment" {
 resource "pingone_notification_policy" "user" {
   environment_id = pingone_environment.my_environment.id
 
-  name = "User Quota SMS, Voice and Email"
+  name = "User Quota SMS Voice and Email"
 
   quota {
     type             = "USER"