+++ internal/service/sso/resource_population.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 	"net/http"
+	"regexp"
 
 	"github.com/hashicorp/terraform-plugin-framework-validators/stringvalidator"
 	"github.com/hashicorp/terraform-plugin-framework/diag"
@@ -13,13 +14,17 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/schema/validator"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/patrickcping/pingone-go-sdk-v2/management"
+	client "github.com/pingidentity/terraform-provider-pingone/internal/client"
 	"github.com/pingidentity/terraform-provider-pingone/internal/framework"
 	"github.com/pingidentity/terraform-provider-pingone/internal/sdk"
 	"github.com/pingidentity/terraform-provider-pingone/internal/verify"
 )
 
 // Types
-type PopulationResource serviceClientType
+type PopulationResource struct {
+	serviceClientType
+	options client.GlobalOptions
+}
 
 type PopulationResourceModel struct {
 	Id               types.String `tfsdk:"id"`
@@ -112,6 +117,11 @@ func (r *PopulationResource) Configure(ctx context.Context, req resource.Configu
 		)
 		return
 	}
+
+	if resourceConfig.Client.GlobalOptions != nil {
+		r.options = *resourceConfig.Client.GlobalOptions
+	}
+
 }
 
 func (r *PopulationResource) Create(ctx context.Context, req resource.CreateRequest, resp *resource.CreateResponse) {
@@ -255,6 +265,20 @@ func (r *PopulationResource) Delete(ctx context.Context, req resource.DeleteRequ
 		return
 	}
 
+	hasUsersAssigned, d := r.hasUsersAssigned(ctx, data.EnvironmentId.ValueString(), data.Id.ValueString())
+	resp.Diagnostics.Append(d...)
+	if resp.Diagnostics.HasError() {
+		return
+	}
+
+	if hasUsersAssigned {
+		d := r.checkControlsAndDeletePopulationUsers(ctx, data.EnvironmentId.ValueString(), data.Id.ValueString())
+		resp.Diagnostics.Append(d...)
+		if resp.Diagnostics.HasError() {
+			return
+		}
+	}
+
 	// Run the API call
 	resp.Diagnostics.Append(framework.ParseResponse(
 		ctx,
@@ -346,6 +370,152 @@ func (p *PopulationResourceModel) toState(apiObject *management.Population) diag
 		p.PasswordPolicyId = framework.StringOkToTF(v.GetIdOk())
 	} else {
 		p.PasswordPolicyId = types.StringNull()
+	}
+
+	return diags
+}
+
+func (r *PopulationResource) hasUsersAssigned(ctx context.Context, environmentID, populationID string) (bool, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	users, d := r.readUsers(ctx, environmentID, populationID)
+	diags.Append(d...)
+	if diags.HasError() {
+		return false, diags
+	}
+
+	if len(users) > 0 {
+		return true, diags
+	}
+
+	return false, diags
+}
+
+func (r *PopulationResource) readUsers(ctx context.Context, environmentID, populationID string) ([]management.User, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	if m, err := regexp.MatchString(verify.P1ResourceIDRegexpFullString.String(), populationID); err == nil && m {
+
+		scimFilter := fmt.Sprintf(`population.id eq "%s"`, populationID)
+
+		// Run the API call
+		var entityArray *management.EntityArray
+		diags.Append(framework.ParseResponse(
+			ctx,
+
+			func() (any, *http.Response, error) {
+				fO, fR, fErr := r.Client.ManagementAPIClient.UsersApi.ReadAllUsers(ctx, environmentID).Filter(scimFilter).Execute()
+				return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, fO, fR, fErr)
+			},
+			"ReadAllUsers",
+			framework.DefaultCustomError,
+			sdk.DefaultCreateReadRetryable,
+			&entityArray,
+		)...)
+
+		if diags.HasError() {
+			return nil, diags
+		}
+
+		return entityArray.Embedded.GetUsers(), nil
+	}
+
+	if r.options.Population.ContainsUsersForceDelete {
+		diags.AddError(
+			"Data protection notice",
+			fmt.Sprintf("For data protection reasons, it could not be determined whether users exist in the population %[2]s in environment %[1]s. Any users in this population will not be deleted.", environmentID, populationID),
+		)
+	}
+
+	return nil, diags
+}
+
+func (r *PopulationResource) checkEnvironmentControls(ctx context.Context, environmentID, populationID string) (bool, diag.Diagnostics) {
+	var diags diag.Diagnostics
+
+	if r.options.Population.ContainsUsersForceDelete {
+		// If the environment options are to force delete production types, then return true, because this will delete the population anyway
+		if r.options.Environment.ProductionTypeForceDelete {
+			return true, diags
+		}
+
+		// Check if the environment is a sandbox type.  We'll only delete users in sandbox environments
+		var environmentResponse *management.Environment
+		diags.Append(framework.ParseResponse(
+			ctx,
+
+			func() (any, *http.Response, error) {
+				fO, fR, fErr := r.Client.ManagementAPIClient.EnvironmentsApi.ReadOneEnvironment(ctx, environmentID).Execute()
+				return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, fO, fR, fErr)
+			},
+			"ReadOneEnvironment-DeletePopulation",
+			framework.DefaultCustomError,
+			nil,
+			&environmentResponse,
+		)...)
+		if diags.HasError() {
+			return false, diags
+		}
+
+		if v, ok := environmentResponse.GetTypeOk(); ok && *v == management.ENUMENVIRONMENTTYPE_SANDBOX {
+			return true, diags
+		} else {
+			diags.AddWarning(
+				"Data protection notice",
+				fmt.Sprintf("For data protection reasons, the provider configuration `global_options.population.contains_users_force_delete` has no effect on environment ID %[1]s as it has a type set to `PRODUCTION`.  Users in this population will not be deleted.\n"+
+					"If you wish to force delete population %[2]s in environment %[1]s, please set the environment type to \"SANDBOX\" or set the `global_options.environment.production_type_force_delete` provider setting to `true`.", environmentID, populationID),
+			)
+		}
+	}
+
+	return false, diags
+}
+
+func (r *PopulationResource) checkControlsAndDeletePopulationUsers(ctx context.Context, environmentID, populationID string) diag.Diagnostics {
+	var diags diag.Diagnostics
+
+	environmentControlsOk, d := r.checkEnvironmentControls(ctx, environmentID, populationID)
+	diags.Append(d...)
+	if diags.HasError() {
+		return diags
+	}
+
+	if environmentControlsOk {
+
+		loopCounter := 1
+		for loopCounter > 0 {
+
+			users, d := r.readUsers(ctx, environmentID, populationID)
+			diags.Append(d...)
+			if diags.HasError() {
+				return diags
+			}
+
+			// DELETE USERS
+			if len(users) == 0 {
+				break
+			} else {
+				for _, user := range users {
+					var entityArray *management.EntityArray
+					diags.Append(framework.ParseResponse(
+						ctx,
+
+						func() (any, *http.Response, error) {
+							fR, fErr := r.Client.ManagementAPIClient.UsersApi.DeleteUser(ctx, environmentID, user.GetId()).Execute()
+							return framework.CheckEnvironmentExistsOnPermissionsError(ctx, r.Client.ManagementAPIClient, environmentID, nil, fR, fErr)
+						},
+						"DeleteUser-DeletePopulation",
+						framework.DefaultCustomError,
+						nil,
+						&entityArray,
+					)...)
+
+					if diags.HasError() {
+						return diags
+					}
+				}
+			}
+		}
 	}
 
 	return diags